reorganizing the house
