+++ srcs/hooks/key_hooks.c
@@ -1,0 +1,45 @@
+#include "cub.h"
+
+int	key_up(int keycode, t_cub *cub)
+{
+	if (keycode == W_KEY)
+		cub->keys.w = 0;
+	if (keycode == A_KEY)
+		cub->keys.a = 0;
+	if (keycode == S_KEY)
+		cub->keys.s = 0;
+	if (keycode == D_KEY)
+		cub->keys.d = 0;
+	if (keycode == UP_KEY)
+		cub->keys.up = 0;
+	if (keycode == DOWN_KEY)
+		cub->keys.down = 0;
+	if (keycode == LEFT_KEY)
+		cub->keys.left = 0;
+	if (keycode == RIGHT_KEY)
+		cub->keys.right = 0;
+	return (0);
+}
+
+int	key_down(int keycode, t_cub *cub)
+{
+	if (keycode == W_KEY)
+		cub->keys.w = 1;
+	if (keycode == A_KEY)
+		cub->keys.a = 1;
+	if (keycode == S_KEY)
+		cub->keys.s = 1;
+	if (keycode == D_KEY)
+		cub->keys.d = 1;
+	if (keycode == UP_KEY)
+		cub->keys.up = 1;
+	if (keycode == DOWN_KEY)
+		cub->keys.down = 1;
+	if (keycode == LEFT_KEY)
+		cub->keys.left = 1;
+	if (keycode == RIGHT_KEY)
+		cub->keys.right = 1;
+	if (keycode == ESC_KEY)
+		graceful_exit(cub);
+	return (0);
+}
