+++ srcs/render/navigator.c
