+++ srcs/render/render.c
@@ -1,47 +1,9 @@
 #include "cub.h"
 
-int	key_up(int keycode, t_cub *cub)
+void	render(t_cub *cub)
 {
-	if (keycode == W_KEY)
-		cub->keys.w = 0;
-	if (keycode == A_KEY)
-		cub->keys.a = 0;
-	if (keycode == S_KEY)
-		cub->keys.s = 0;
-	if (keycode == D_KEY)
-		cub->keys.d = 0;
-	if (keycode == UP_KEY)
-		cub->keys.up = 0;
-	if (keycode == DOWN_KEY)
-		cub->keys.down = 0;
-	if (keycode == LEFT_KEY)
-		cub->keys.left = 0;
-	if (keycode == RIGHT_KEY)
-		cub->keys.right = 0;
-	return (0);
-}
-
-int	key_down(int keycode, t_cub *cub)
-{
-	if (keycode == W_KEY)
-		cub->keys.w = 1;
-	if (keycode == A_KEY)
-		cub->keys.a = 1;
-	if (keycode == S_KEY)
-		cub->keys.s = 1;
-	if (keycode == D_KEY)
-		cub->keys.d = 1;
-	if (keycode == UP_KEY)
-		cub->keys.up = 1;
-	if (keycode == DOWN_KEY)
-		cub->keys.down = 1;
-	if (keycode == LEFT_KEY)
-		cub->keys.left = 1;
-	if (keycode == RIGHT_KEY)
-		cub->keys.right = 1;
-	if (keycode == ESC_KEY)
-		graceful_exit(cub);
-	return (0);
+	raycast(cub);
+	draw_navigator(cub);
 }
 
 /*
@@ -77,6 +39,6 @@ int	update_display(t_cub *cub)
 	if (cub->keys.w || cub->keys.a || cub->keys.s || cub->keys.d
 		|| cub->keys.up || cub->keys.down || cub->keys.left
 		|| cub->keys.right)
-		raycast(cub);
+		render(cub);
 	return (0);
 }
