+++ srcs/render/texture_color.c
