+++ srcs/render/textures.c
